requirements.txt
# Основные зависимости для B2B AutoDesk

# GUI Framework
PyQt5>=5.15.0
# Альтернатива: PySide6>=6.0.0

# База данных
psycopg2-binary>=2.9.0

# Конфигурация
python-dotenv>=1.0.0

# Логирование
loguru>=0.7.0

# Нечеткий поиск (fuzzy search) для обработки опечаток
rapidfuzz>=3.0.0
# Альтернатива (если rapidfuzz недоступен):
# fuzzywuzzy>=0.18.0
# python-Levenshtein>=0.12.0

# Работа с документами
requests>=2.32.0
openpyxl>=3.1.2
xlrd==1.2.0  # Для чтения старых .xls файлов (версия 2.0+ не поддерживает старые .xls)
xlwt>=1.3.0  # Для создания тестовых .xls файлов
pandas>=2.0.0  # Резервный способ открытия Excel файлов
python-docx>=1.0.0  # Для парсинга Word документов (.docx)
pywin32>=305; platform_system == "Windows"  # Для Word COM объектов в Windows (.doc файлы)
PyPDF2>=3.0.0  # Для парсинга PDF файлов
pdfplumber>=0.10.0  # Альтернативный парсер PDF (лучшее качество)
pytesseract>=0.3.10  # OCR для отсканированных PDF
Pillow>=10.0.0  # Для работы с изображениями в OCR
pdf2image>=1.16.0  # Для конвертации PDF в изображения для OCR
reportlab>=4.0.0  # Для создания тестовых PDF файлов
rarfile>=4.2
py7zr>=0.21.0
psutil>=5.9.0  # Для определения процессов, блокирующих файлы

